+++ verik-compiler/src/main/kotlin/io/verik/compiler/main/VerikCompilerReport.kt
@@ -7,7 +7,7 @@ package io.verik.compiler.main
 /**
  * Report data structure that stores some statistics.
  */
-class VerikReport {
+class VerikCompilerReport {
 
     var entryPoints: List<String> = listOf()
     var counts: List<Pair<String, Int>> = listOf()