+++ verik-plugin/src/main/kotlin/io/verik/plugin/config/TargetConfig.kt
@@ -1,0 +1,17 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.verik.plugin.config
+
+import java.nio.file.Path
+
+/**
+ * Abstract target configuration.
+ */
+abstract class TargetConfig {
+
+    abstract val projectConfig: ProjectConfig
+    abstract val name: String
+    abstract val buildDir: Path
+}