+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/VerikDomainObject.kt
@@ -1,0 +1,19 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.verik.plugin.domain
+
+/**
+ * Interface for domain objects that configure the toolchain.
+ */
+interface VerikDomainObject {
+
+    var maxErrorCount: Int
+    var debug: Boolean
+
+    fun import(configure: VerikImportDomainObject.() -> Unit)
+    fun compile(configure: VerikCompileDomainObject.() -> Unit)
+    fun dsim(configure: DsimTargetDomainObject.() -> Unit)
+    fun vivado(configure: VivadoTargetDomainObject.() -> Unit)
+}