+++ CHANGELOG.md
@@ -7,6 +7,14 @@
 ### Compiler
 - Document compiler classes.
 
+### Plugin
+- Rewrite plugin extension domain object classes.
+- Move compiler to `verikCompile` task.
+- Add `verik` task that generates build scripts for downstream tools.
+- Move compiler and importer build directories to `build/compile` and `build/import`.
+- Add the Metrics dsim generation target.
+- Add the Xilinx vivado generation target.
+
 ## [0.1.15]
 ### Compiler
 - Support `Double` data type.